+++ app/Http/Controllers/Cms/UserController.php
@@ -6,10 +6,14 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
 use App\Http\Controllers\Controller;
+use App\Traits\ImageableTrait;
+use App\Enums\ImageTypeEnum;
 use App\Models\User;
+use App\Models\Image;
 
 class UserController extends Controller
 {
+    use ImageableTrait;
 
     public function __construct()
     {
@@ -25,21 +29,18 @@ class UserController extends Controller
     {
         $users = User::orderBy('created_at')->get();
         
-        return response()->json([
-            'status' => 200,
-            'meta' => [
-                'count' => count($users),
-                'entityType' => 'users',
-                'headers' => [
-                    [ 'entityKey' => 'name', 'type' => 'text', 'value' => 'Név'], 
-                    [ 'entityKey' => 'username', 'type' => 'text', 'value' => 'Felhasználónév'], 
-                    [ 'entityKey' => 'email', 'type' => 'text', 'value' => 'Email']
-                ],
-                'key' => 'id', 
-                'value' => 'name'
-            ],
-            'data' => $users,
-        ]);
+        return response()->json($this->handleListResponseSuccess(
+            count($users), 
+            'users', 
+            $users, 
+            [
+                [ 'entityKey' => 'name', 'type' => 'text', 'value' => 'Név'], 
+                [ 'entityKey' => 'username', 'type' => 'text', 'value' => 'Felhasználónév'], 
+                [ 'entityKey' => 'email', 'type' => 'text', 'value' => 'Email']
+            ], 
+            'id', 
+            'name'
+        ));
     }
 
     /**
@@ -56,14 +57,12 @@ class UserController extends Controller
                 'email' => 'required|string|email|max:200|unique:users',
                 'username' => 'required|string|max:100|unique:users',
                 'password' => 'required|string|min:6',
-                'confirmPassword' => 'required|string|same:password'
+                'confirmPassword' => 'required|string|same:password',
+                'avatar' => 'required|file'
             ]);
         } catch(ValidationException $ve) {
 
-            return response()->json([
-                'status' => 400,
-                'error' => $ve->errors()
-            ], 400);
+            return response()->json($this->handleResponseError(400, $ve->errors()), 400);
         }
 
         $user = User::create([
@@ -73,14 +72,17 @@ class UserController extends Controller
             'password' => Hash::make($request->password),
         ]);
 
-        return response()->json([
-            'status' => 200,
-            'meta' => [
-                'count' => 1,
-                'entityType' => 'users',
-            ],
-            'data' => $user
+        $path = $this->handleImage($request, $user, 'avatar');
+        
+        $image = Image::create([
+            'type' => ImageTypeEnum::AVATAR,
+            'imageabble_type' => get_class($user),
+            'imageable_id' => $user->id,
+            'path' => $path,
+            'title' => $request->username." avatar",
         ]);
+
+        return response()->json($this->handleResponseSuccess(1, 'users', $user));
     }
 
     /**
@@ -92,22 +94,10 @@ class UserController extends Controller
     public function show($id)
     {
         $user = User::find($id);
-        
-        if (!$user) {
-            return response()->json([
-                'status' => 404,
-                'error' => ['user' => 'User not found.']
-            ], 404);
-        }
-        
-        return response()->json([
-            'status' => 200,
-            'meta' => [
-                'count' => 1,
-                'entityType' => 'users',
-            ],
-            'data' => $user,
-        ]);
+
+        $response = $this->handleEntityExist($user, 'users', 'user');
+
+        return response()->json($response, $response['status']);
     }
 
     /**
@@ -121,11 +111,10 @@ class UserController extends Controller
     {
         $user = User::find($id);
 
-        if (!$user) {
-            return response()->json([
-                'status' => 404,
-                'error' => ['user' => 'User not found.']
-            ], 404);
+        $resp = $this->handleEntityExist($user, 'users', 'user');
+        if ($resp['status'] != 200 ) {
+            
+            return response()->json($resp, $resp['status']);
         }
 
         if ($request->get('password')) {
@@ -135,11 +124,8 @@ class UserController extends Controller
                     'confirmPassword' => 'required|string|same:password'
                 ]);
             } catch(ValidationException $ve) {
-    
-                return response()->json([
-                    'status' => 400,
-                    'error' => $ve->errors()
-                ], 400);
+
+                return response()->json($this->handleResponseError(400, $ve->errors()), 400);
             }
     
             $user->password = Hash::make($request->password);
@@ -168,26 +154,30 @@ class UserController extends Controller
             $validationRules['name'] = 'required|string|max:200';
         }
 
+        if ($request->has('avatar') && $request->hasFile('avatar')) {
+            $validationRules['avatar'] = "required|file";
+
+            $path = $this->handleImage($request, $user, 'avatar');
+        
+            $image = Image::create([
+                'type' => ImageTypeEnum::AVATAR,
+                'imageabble_type' => get_class($user),
+                'imageable_id' => $user->id,
+                'path' => $path,
+                'title' => $user->username." avatar",
+            ]);
+        }
+
         try {
             $request->validate($validationRules);
         } catch(ValidationException $ve) {
 
-            return response()->json([
-                'status' => 400,
-                'error' => $ve->errors()
-            ], 400);
+            return response()->json($this->handleResponseError(400, $ve->errors()), 400);
         }
 
         $user->save();
 
-        return response()->json([
-            'status' => 200,
-            'meta' => [
-                'count' => 1,
-                'entityType' => 'users',
-            ],
-            'data' => $user
-        ]);
+        return response()->json($this->handleResponseSuccess(1, 'users', $user));
     }
 
     /**
@@ -200,16 +190,15 @@ class UserController extends Controller
     {
         $user = User::find($id);
 
+        $resp = $this->handleEntityExist($user, 'users', 'user');
+        if ($resp['status'] != 200 ) {
+            
+            return response()->json($resp, $resp['status']);
+        }
+
         $user->delete();
 
-        return response()->json([
-            'status' => 200,
-            'meta' => [
-                'count' => 1,
-                'entityType' => 'users',
-            ],
-            'data' => null
-        ]);
+        return response()->json($this->handleResponseSuccess(0, 'users', null));
     }
 
     public function unique(string $key, string $value = '')
